C++ Tsfile V4 read (#409)
* save work

* save work

* fix for include circle

* save work

* save work

* save work

* save work

* save work

* fix memleak

* fix mem leak

* save work

* remove some files

* save work

* save

* fix complie error

* save work

* fix MeasurementSchema codec

* fix StringStatistic

* fix for merge

* save work

* save work

* save work

* save work

* fix writer interface

* fix writer interface

* complete test and fix some bug

* complete ut

* save work

* update gitignore and fix sorted_array

* fix for memleak

* fix memory leak of: DeviceQueryTask

* fix memory leak of MetdataIndexNode

* remove columnDesc

* fix ut bug

* add metadata interface

* fix v3 write_tablet_aligned

* fix v4 write_table(aligned)

* add debug macro

* multiple fixed regarding using aligned series in table interfaces

* support aligned table read

* fix some problem

* change index to start from 1

* fix warnings

* fix comments

* fix test

* add static assert

* fix time chunk writer leak

* fix merge problem

* fix duplicatedly written time

* add print

* add print

* fix assertion

---------

+++ cpp/src/common/container/sorted_array.h
@@ -246,7 +246,7 @@ class SortedArray {
         int64_t left = 0;
         int64_t right = size_ - 1;
         while (left <= right) {
-            size_t mid = (left + right) / 2;
+            int64_t mid = (left + right) / 2;
             ValueType mid_val = array_[mid];
             if (mid_val == target_val) {
                 found = true;
